Fix for issue #10
This commit resolves issue #10 by silently failing if the LiveCode
Server process can not be spawned. In the process the dependency on
atom-linter was factored out as it was necessary to re-implement
helper.exec(). Rather than reject the promise as is the normal procedure
in the event of an error here I have chosen to resolve with an empty
output because it means you can fix the LiveCode Server process name
and it will start linting in the same session.