Improve support for if...then...else control structure
This commit adds support for:

if condition then something1 else something2

It also adds support for the following with a quirk on the else.
Atom seems to decrease the indent when it hits a match but
doesn't increase it again when it no longer matches.
This means when you are typing else it will step back at the
end of the word and if you keep typing it stays at the wrong
indent instead of popping back as it should. Using the menu to
reset the indents does fix it though. I guess this is
reportable on Atom...

if condition
then something1
else something2